fix: make "use `set_option diagnostics true" message conditional on current setting
It is confusing that the message suggesting to use the `diagnostics` option is given even when the option is already set. This PR makes use of lazy message data to make the message contingent on the option being false.

It also tones down the promise that there is any diagonostic information available, since sometimes there is nothing to report.

Suggested by Johan Commelin.