feat: make `structure` command encode (some) optional parameters in the constructor
The intended way to construct structure values is structure instance notation (`{ x := 1, y := 2 }`). However, for simple structures it is sometimes convenient to use the constructor directly, given that a structure is conceptually a more featureful `inductive` command.

It's not possible to encode default values as optional parameters in all cases due to the ability to override default values for fields in parents using fields that are "after" that field. However, it is safe to encode default values as optional parameters if that field is new to the structure being defined. Safe in the sense that elaboration via the constructor will not diverge from elaboration via structure instance notation, at least for that particular field.

Closes #4736