fix: prevent `end` from closing sections and namespaces simultaneously
Adds a syntax ref to `Command.Scope` to keep track of what is responsible for adding it. In the future, this syntax ref could be used to include links to `section`/`namespace` commands in `end`'s errors. For now, these are used by `end` to determine whether scopes were added by `section` or `namespace` for further validation.

Adds tests for the `end` command.

Removes the unused `Lean.Elab.Command.withNamespace`.