feat: make loose docstrings into elaboration errors rather than parse errors
Now loose docstrings will not yield a confusing parser error and instead log an "unexpected docstring" elaboration error and continue processing.

Furthermore, if the docstring is preceding the `in` command combinator, it will add a hint to the error:
```
unexpected doc string

Hint: the docstring must come after '... in', immediately before the declaration that accepts a docstring.
```

Closes #3135