fix: reduce default max depth for ext tactic
The prior default of 1000000 could not be achieved in practice,
because the stack would overflow after around 5000 recursive
invocations. This meant that a poorly-chosen @[ext] lemma could crash
Lean.

Talking to Mathlib users, it seems that 10 would be a
very large number in practice, so a default limit of 100 should not
change successful uses. But it does make it much easier to diagnose
and recover from poor choices of @[ext] lemmas.