feat: create CACHEDIR.TAG file in .lake directory
Create CACHEDIR.TAG file in .lake directory when it is first created.
The file does not get recreated if removed if .lake directory is not
removed, too. While technically against the spec, this is how Rust does
it, too. Moreover, this simplifies the implementation a lot.

This change also introduces a test to check that the file gets cretead
but not recreated. This test needs to change e.g. if we decide that
CACHEDIR.TAG should always be recreated.